LayoutTests/imported/w3c/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/embedded-and-images-presentational-hints-ascii-case-insensitive-expected.txt

        

PASS keyword absbottom 
FAIL keyword abscenter assert_equals: lowercase valid expected "middle" but got "baseline"
PASS keyword absmiddle 
